requirements.txt
amqp==5.1.1
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.6
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==2.1
django-celery-results==2.3.1
djangorestframework==3.9.0
idna==3.3
importlib-metadata==4.11.3
kombu==5.2.4
mysqlclient==2.1.0
prompt-toolkit==3.0.29
pytz==2022.1
requests==2.27.1
six==1.16.0
typing_extensions==4.2.0
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
zipp==3.8.0
